requirements.txt
line-bot-sdk
flask
requests
bs4
python-dotenv
gunicorn

